requirements.txt
yfinance==1.2.0
pandas==2.3.3
pandas-datareader==0.10.0
numpy==2.2.3
psycopg2-binary==2.9.10
seaborn==0.13.2
python-dotenv==1.0.0
requests==2.32.5
pytz==2025.2   
sqlalchemy==2.0.36
fredapi==0.4.3
